requirements.txt
blessings==1.7
bpython==0.18
certifi==2019.6.16
chardet==3.0.4
curtsies==0.3.0
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-extensions==2.1.9
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.9.4
greenlet==0.4.15
idna==2.8
oauthlib==3.0.2
Pillow==6.0.0
Pygments==2.4.2
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typing==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
